+++ src/HotChocolate/Data/src/Overhaul/ExpressionNodes/Sealed/SealedExpressionNode.cs
@@ -1,12 +1,10 @@
-using System.Collections.Generic;
-
 namespace HotChocolate.Data.ExpressionNodes;
 
 internal readonly record struct SealedExpressionNode(
     SealedScope? Scope,
     IExpressionFactory ExpressionFactory,
     ReadOnlyStructuralDependencies Dependencies,
-    IReadOnlyList<Identifier> Children)
+    Identifier[] Children)
 {
     public bool IsInitialized => ExpressionFactory is not null;
 }
